fix(ma): security hardening + consortium state rules + extraction pipeline improvements
- Add check_client_deal_access() to marketing log create/update/delete endpoints
- Implement consortium status transition rules (TAPPING→CONFIRMED→DROPPED)
- Block SOLE_BUYER from consortium participation with 422 error
- Add blob storage download + temp file parsing for document extraction
- Add 5-min pipeline timeout with graceful FAILED marking
- Add ensure_initialized() to BlobStorageClient for Celery worker safety
- Fix Decimal types in DartFinancialSummaryOut (float→Decimal)
- Fix Excel export IOI/LOI values (float→int for KRW amounts)
- Add logger.warning to integration error handlers
- Apply design system tokens to ConsortiumPanel (slate-*→semantic)
- Add accessibility: htmlFor/id, aria-expanded, scope="col", aria-label
- Add EngagementDocUpload polling timeout (5min) with user-facing error
- Add consortium tests: sole_buyer, state transition, dropped terminal, equity boundary

+++ deal-mgmt/app/schemas/marketing_log.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 
 import uuid
 from datetime import datetime
+from decimal import Decimal
 
 from pydantic import BaseModel, ConfigDict, Field
 
@@ -46,8 +47,8 @@ class BuyerStageSummary(BaseModel):
 class DartFinancialSummaryOut(BaseModel):
     """DART 재무 요약 응답 스키마."""
 
-    revenue: float | None = None
-    operating_profit: float | None = None
-    net_income: float | None = None
-    debt_ratio: float | None = None
+    revenue: Decimal | None = None
+    operating_profit: Decimal | None = None
+    net_income: Decimal | None = None
+    debt_ratio: Decimal | None = None
     fiscal_year: str | None = None
